21/11 19:42

+++ BL/BO/Product.cs
@@ -10,7 +10,7 @@ namespace BO
     {
         public int ID { get; set; }
         public string Name { get; set; }
-        public double Price { get; set; }
+        public float Price { get; set; }
         public BO.Categories Category { get; set; }
         public int InStock { get; set; }
